% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatology.R
\name{ERA5.climatology.get.file.name}
\alias{ERA5.climatology.get.file.name}
\title{ERA5 get file name for hourly climatological data}
\usage{
ERA5.climatology.get.file.name(variable, month)
}
\arguments{
\item{variable}{'prmsl', 'prate', 'air.2m', 'uwnd.10m' or 'vwnd.10m' - or any supported variable}
}
\value{
File containing the requested data
}
\description{
Get the file name for selected variable and date (hourly data)
}
