% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fetch.data.R
\name{ERA5.fetch.analysis.data.for.month}
\alias{ERA5.fetch.analysis.data.for.month}
\title{Get data for one analysis variable}
\usage{
ERA5.fetch.analysis.data.for.month(var, year, month, stream = "oper")
}
\arguments{
\item{var}{- 20CR variable name}

\item{stream}{- 'ensda' for the lower resolution ensemble, otherwise 'oper' (default)}
}
\description{
I'm packaging data by the month, a compromise
  between speed, space, & simplicity.
}
\details{
Won't replace data already on disc - just
  returns immediately if the result file is already there.
}
