% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fetch.data.R
\name{ERA5.fetch.data.for.month}
\alias{ERA5.fetch.data.for.month}
\title{Get the data for one variable for one month}
\usage{
ERA5.fetch.data.for.month(var, year, month, stream = "oper")
}
\arguments{
\item{var}{- 20CR variable name}

\item{stream}{- 'ensda' for the lower resolution ensemble, otherwise 'oper' (default)}
}
\description{
Use 20CR variable names, store the data on $SCRATCH
}
