% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GSDF.ERA5.R
\name{ERA5.translate.for.files}
\alias{ERA5.translate.for.files}
\title{Translate to ERA5 variable names as used in files}
\usage{
ERA5.translate.for.files(var)
}
\arguments{
\item{var}{- 20CR variable name}
}
\value{
name used for ERA5 file names
}
\description{
I'm standardising on 20CR variable names - map to ERA5 choices
}
\details{
ERA5 uses different names for the files and the variable in the file
 this function maps to ERA5 file names.
}

