% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.ERA5.R
\name{ERA5.translate.for.variable.names}
\alias{ERA5.translate.for.variable.names}
\title{Translate to ERA5 variable names in the .nc files}
\usage{
ERA5.translate.for.variable.names(var)
}
\arguments{
\item{var}{- 20CR variable name}
}
\value{
name used for ERA5 variable names
}
\description{
I'm standardising on 20CR variable names - map to ERA5 choices
}
\details{
ERA5 uses different names for the files and the variable in the file
 this function maps to ERA5 variable names.
}
